Move remaining CLI logic into `capsula-orchestration` crate
Extract vault queries (list_runs, find_run_dir, find_run_dir_by_name),
push logic (push_single_run), and config resolution (resolve_vault_path,
resolve_server_url) from capsula-cli into capsula-orchestration.

The CLI is now a thin shell over capsula-orchestration, making the
orchestration logic reusable by a future PyO3 Python wrapper.
