+++ crates/capsula-cli/src/main.rs
@@ -8,14 +8,15 @@ use anyhow::{Context, Result};
 use capsula_config::CapsulaConfig;
 use capsula_core::hook::{PostRun, PreRun};
 use capsula_core::run::Run;
+use capsula_orchestration::push::push_single_run;
+use capsula_orchestration::resolve::{resolve_server_url, resolve_vault_path};
 use capsula_orchestration::run::{create_and_setup_run, run_post_hooks, run_pre_hooks};
+use capsula_orchestration::vault::{find_run_dir, find_run_dir_by_name, list_runs};
 use chrono::DateTime;
 use clap::{Parser, Subcommand};
-use serde::Deserialize;
 use std::path::PathBuf;
 use tracing::{debug, error, info, warn};
 use tracing_subscriber::{EnvFilter, fmt};
-use ulid::Ulid;
 
 const VERSION: &str = if env!("GIT_HASH").is_empty() {
     env!("CARGO_PKG_VERSION")
@@ -95,406 +96,21 @@ enum VaultsCommands {
     },
 }
 
-/// Resolve the vault path with priority: CLI argument > environment variable > config file.
-///
-/// This function is called after dotenv loading, so environment variables from the dotenv file
-/// are available. This is important because Clap's `env` attribute reads environment variables
-/// at parse time, before the dotenv file is loaded.
-fn resolve_vault_path(
-    cli_vault_path: Option<PathBuf>,
-    config_vault_path: &std::path::Path,
-    project_root: &std::path::Path,
-) -> PathBuf {
-    // Priority 1: CLI argument
-    if let Some(path) = cli_vault_path {
-        debug!("Using vault path from CLI argument: {}", path.display());
-        return if path.is_absolute() {
-            path
-        } else {
-            project_root.join(path)
-        };
-    }
-
-    // Priority 2: Environment variable (checked after dotenv loading)
-    if let Ok(env_path) = std::env::var("CAPSULA_VAULT_PATH") {
-        let path = PathBuf::from(&env_path);
-        debug!(
-            "Using vault path from CAPSULA_VAULT_PATH env var: {}",
-            path.display()
-        );
-        return if path.is_absolute() {
-            path
-        } else {
-            project_root.join(path)
-        };
-    }
-
-    // Priority 3: Config file
-    debug!(
-        "Using vault path from config file: {}",
-        config_vault_path.display()
-    );
-    if config_vault_path.is_absolute() {
-        config_vault_path.to_path_buf()
-    } else {
-        project_root.join(config_vault_path)
-    }
-}
-
-/// Resolve the server URL with priority: CLI argument > environment variable > config file.
-///
-/// This function is called after dotenv loading, so environment variables from the dotenv file
-/// are available. This is important because Clap's `env` attribute reads environment variables
-/// at parse time, before the dotenv file is loaded.
-fn resolve_server_url(cli_server: Option<String>, config_server: Option<&str>) -> Option<String> {
-    // Priority 1: CLI argument
-    if let Some(server) = cli_server {
-        debug!("Using server URL from CLI argument: {}", server);
-        return Some(server);
-    }
-
-    // Priority 2: Environment variable (checked after dotenv loading)
-    if let Ok(env_server) = std::env::var("CAPSULA_SERVER_URL") {
-        debug!(
-            "Using server URL from CAPSULA_SERVER_URL env var: {}",
-            env_server
-        );
-        return Some(env_server);
-    }
-
-    // Priority 3: Config file
-    if let Some(server) = config_server {
-        debug!("Using server URL from config file: {}", server);
-        return Some(server.to_string());
-    }
-
-    None
-}
-
-fn create_pre_run_hook_registry() -> capsula_registry::HookRegistry<PreRun> {
-    // Use the standard registry with all built-in hook types
-    capsula_registry::standard_pre_run_hook_registry()
-}
-
-fn create_post_run_hook_registry() -> capsula_registry::HookRegistry<PostRun> {
-    // Use the standard registry with all built-in hook types
-    capsula_registry::standard_post_run_hook_registry()
-}
-
-#[derive(Debug, Deserialize)]
-struct RunMetadata {
-    id: Ulid,
-    name: String,
-    command: Vec<String>,
-    timestamp: String,
-}
-
-fn list_runs(vault_dir: &std::path::Path) -> Result<Vec<RunMetadata>> {
-    let mut runs = Vec::new();
-
-    // Check if vault directory exists
-    if !vault_dir.exists() {
-        return Ok(runs);
-    }
-
-    // Iterate through date directories
-    for date_entry in std::fs::read_dir(vault_dir)
-        .with_context(|| format!("Failed to read vault directory: {}", vault_dir.display()))?
-    {
-        let date_entry = date_entry?;
-        let date_path = date_entry.path();
-
-        // Skip if not a directory or is .gitignore
-        if !date_path.is_dir() {
-            continue;
-        }
-
-        // Iterate through run directories within each date
-        for run_entry in std::fs::read_dir(&date_path)
-            .with_context(|| format!("Failed to read date directory: {}", date_path.display()))?
-        {
-            let run_entry = run_entry?;
-            let run_path = run_entry.path();
-
-            // Skip if not a directory
-            if !run_path.is_dir() {
-                continue;
-            }
-
-            // Look for _capsula/metadata.json
-            let metadata_path = run_path.join("_capsula").join("metadata.json");
-            if metadata_path.exists() {
-                // Read and parse metadata
-                match std::fs::read_to_string(&metadata_path) {
-                    Ok(content) => match serde_json::from_str::<RunMetadata>(&content) {
-                        Ok(metadata) => runs.push(metadata),
-                        Err(e) => {
-                            warn!(
-                                "Failed to parse metadata from {}: {}",
-                                metadata_path.display(),
-                                e
-                            );
-                        }
-                    },
-                    Err(e) => {
-                        warn!(
-                            "Failed to read metadata from {}: {}",
-                            metadata_path.display(),
-                            e
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    // Sort by timestamp (newest first)
-    runs.sort_by(|a, b| {
-        // Parse timestamps and compare
-        let time_a = DateTime::parse_from_rfc3339(&a.timestamp).ok();
-        let time_b = DateTime::parse_from_rfc3339(&b.timestamp).ok();
-
-        match (time_a, time_b) {
-            (Some(a), Some(b)) => b.cmp(&a), // Reverse order for newest first
-            (Some(_), None) => std::cmp::Ordering::Less,
-            (None, Some(_)) => std::cmp::Ordering::Greater,
-            (None, None) => std::cmp::Ordering::Equal,
-        }
-    });
-
-    Ok(runs)
-}
-
-fn push_single_run(
-    run_dir: &std::path::Path,
-    vault_name: &str,
-    server_url: &str,
-    client: &capsula_client::CapsulaClient,
-) -> Result<()> {
-    let capsula_dir = run_dir.join("_capsula");
-
-    // Read metadata
-    let metadata_path = capsula_dir.join("metadata.json");
-    let metadata_content = std::fs::read_to_string(&metadata_path)
-        .with_context(|| format!("Failed to read metadata from {}", metadata_path.display()))?;
-    let metadata: serde_json::Value = serde_json::from_str(&metadata_content)?;
-
-    let run_name = metadata["name"].as_str().unwrap_or("unknown");
-    info!("Pushing run: {}", run_name);
-
-    // Read pre-run and post-run hooks
-    let pre_run_path = capsula_dir.join("pre-run.json");
-    let pre_run_hooks = if pre_run_path.exists() {
-        let content = std::fs::read_to_string(&pre_run_path)?;
-        Some(serde_json::from_str::<Vec<serde_json::Value>>(&content)?)
-    } else {
-        None
-    };
-
-    let post_run_path = capsula_dir.join("post-run.json");
-    let post_run_hooks = if post_run_path.exists() {
-        let content = std::fs::read_to_string(&post_run_path)?;
-        Some(serde_json::from_str::<Vec<serde_json::Value>>(&content)?)
-    } else {
-        None
-    };
-
-    // Read command output
-    let command_json_path = capsula_dir.join("command.json");
-    let command_output = if command_json_path.exists() {
-        let content = std::fs::read_to_string(&command_json_path)?;
-        serde_json::from_str::<serde_json::Value>(&content)?
-    } else {
-        serde_json::json!({})
-    };
-
-    // Convert duration object to milliseconds
-    let duration_ms = command_output.get("duration").and_then(|d| {
-        let secs = d.get("secs")?.as_u64()?;
-        let nanos = d.get("nanos")?.as_u64()?;
-        Some((secs * 1000) + (nanos / 1_000_000))
-    });
-
-    let create_run_payload = serde_json::json!({
-        "id": metadata["id"],
-        "name": metadata["name"],
-        "timestamp": metadata["timestamp"],
-        "command": serde_json::to_string(&metadata["command"])?,
-        "vault": vault_name,
-        "project_root": metadata["project_root"],
-        "exit_code": command_output.get("exit_code"),
-        "duration_ms": duration_ms,
-        "stdout": command_output.get("stdout"),
-        "stderr": command_output.get("stderr"),
-    });
-
-    // Post the run metadata
-    let url = format!("{server_url}/api/v1/runs");
-    let http_client = reqwest::blocking::Client::new();
-    let response = http_client.post(&url).json(&create_run_payload).send()?;
-
-    if !response.status().is_success() {
-        if response.status().as_u16() == 409 {
-            // Return a special error that caller can detect
-            anyhow::bail!("Run already exists: {run_name}");
-        }
-        anyhow::bail!("Failed to create run on server: {}", response.status());
-    }
-
-    // Collect files to upload
-    let mut files = Vec::new();
-    for entry in walkdir::WalkDir::new(run_dir)
-        .into_iter()
-        .filter_entry(|e| e.file_name() != "_capsula")
-    {
-        let entry = entry?;
-        if entry.file_type().is_file() {
-            let local_path = entry.path();
-            let relative_path = local_path.strip_prefix(run_dir)?;
-            files.push((local_path.to_path_buf(), relative_path.to_path_buf()));
-        }
-    }
-
-    // Upload files and hooks
-    if !files.is_empty() || pre_run_hooks.is_some() || post_run_hooks.is_some() {
-        let actual_run_id = metadata["id"]
-            .as_str()
-            .ok_or_else(|| anyhow::anyhow!("Run ID not found in metadata"))?;
-        let response = client.upload_run(actual_run_id, &files, pre_run_hooks, post_run_hooks)?;
-        debug!(
-            "Upload complete: {} files, {} bytes",
-            response.files_processed, response.total_bytes
-        );
-    }
-
-    info!("✓ Pushed: {}", run_name);
-    Ok(())
-}
-
-fn find_run_dir(vault_dir: &std::path::Path, run_id: &str) -> Result<PathBuf> {
-    // Search for the run directory by ID or name
-    // Structure: vault_dir/{YYYY-MM-DD}/{HHMMSS-name}/
-    for entry in walkdir::WalkDir::new(vault_dir)
-        .min_depth(2)
-        .max_depth(2)
-        .into_iter()
-        .filter_entry(|e| e.file_type().is_dir())
-    {
-        let entry = entry?;
-        let capsula_dir = entry.path().join("_capsula");
-        if !capsula_dir.exists() {
-            continue;
-        }
-
-        let metadata_path = capsula_dir.join("metadata.json");
-        if !metadata_path.exists() {
-            continue;
-        }
-
-        let metadata_content = std::fs::read_to_string(&metadata_path)?;
-        let metadata: serde_json::Value = serde_json::from_str(&metadata_content)?;
-
-        // Check if ID matches
-        if let Some(id) = metadata.get("id").and_then(|v| v.as_str())
-            && id == run_id
-        {
-            return Ok(entry.path().to_path_buf());
-        }
-
-        // Check if name matches
-        if let Some(name) = metadata.get("name").and_then(|v| v.as_str())
-            && name == run_id
-        {
-            return Ok(entry.path().to_path_buf());
-        }
-    }
-
-    anyhow::bail!(
-        "Run with ID or name '{}' not found in vault {}",
-        run_id,
-        vault_dir.display()
-    )
-}
-
-fn find_run_dir_by_name(vault_dir: &std::path::Path, run_name: &str) -> Result<PathBuf> {
-    let mut best_match: Option<(PathBuf, Option<DateTime<chrono::FixedOffset>>)> = None;
-    let mut match_count = 0usize;
-
-    for entry in walkdir::WalkDir::new(vault_dir)
-        .min_depth(2)
-        .max_depth(2)
-        .into_iter()
-        .filter_entry(|e| e.file_type().is_dir())
-    {
-        let entry = entry?;
-        let capsula_dir = entry.path().join("_capsula");
-        if !capsula_dir.exists() {
-            continue;
-        }
-
-        let metadata_path = capsula_dir.join("metadata.json");
-        if !metadata_path.exists() {
-            continue;
-        }
-
-        let metadata_content = std::fs::read_to_string(&metadata_path)?;
-        let metadata: RunMetadata = serde_json::from_str(&metadata_content)?;
-
-        if metadata.name != run_name {
-            continue;
-        }
-
-        match_count += 1;
-        let timestamp = DateTime::parse_from_rfc3339(&metadata.timestamp).ok();
-        match best_match {
-            None => {
-                best_match = Some((entry.path().to_path_buf(), timestamp));
-            }
-            Some((_, ref best_timestamp)) => {
-                let is_newer = match (timestamp, best_timestamp) {
-                    (Some(current), Some(best)) => current > *best,
-                    (Some(_), None) => true,
-                    _ => false,
-                };
-                if is_newer {
-                    best_match = Some((entry.path().to_path_buf(), timestamp));
-                }
-            }
-        }
-    }
-
-    if let Some((path, _)) = best_match {
-        if match_count > 1 {
-            warn!(
-                "Found {} runs named '{}'; returning the newest by timestamp.",
-                match_count, run_name
-            );
-        }
-        return Ok(path);
-    }
-
-    anyhow::bail!(
-        "Run with name '{}' not found in vault {}",
-        run_name,
-        vault_dir.display()
-    )
-}
-
 #[expect(
     clippy::too_many_lines,
-    reason = "Remaining length is from config loading and non-run commands (List, Push, Vaults)"
+    reason = "CLI dispatch function; further splitting would hurt readability"
 )]
 fn run() -> Result<()> {
-    // Create the registry with all available hook types
     debug!("Creating hook registries");
-    let pre_run_hook_registry = create_pre_run_hook_registry();
-    let post_run_hook_registry = create_post_run_hook_registry();
+    let pre_run_hook_registry: capsula_registry::HookRegistry<PreRun> =
+        capsula_registry::standard_pre_run_hook_registry();
+    let post_run_hook_registry: capsula_registry::HookRegistry<PostRun> =
+        capsula_registry::standard_post_run_hook_registry();
 
     let cli = Cli::parse();
     let config_file_path = cli.config.unwrap_or_else(|| PathBuf::from("capsula.toml"));
     debug!("Using configuration file: {}", config_file_path.display());
 
-    // Check if the config file exists
     if !config_file_path.exists() {
         anyhow::bail!(
             "Configuration file not found at '{}'
@@ -516,7 +132,6 @@ path = \".\"
         );
     }
 
-    // Canonicalize the config file path first to get an absolute path
     let config_file_path = config_file_path.canonicalize().with_context(|| {
         format!(
             "Failed to resolve configuration file path: {}",
@@ -584,13 +199,11 @@ path = \".\"
             println!("{}", "-".repeat(19 + 2 + 20 + 2 + command_width));
 
             for run in runs {
-                // Parse timestamp for display
                 let timestamp_display = DateTime::parse_from_rfc3339(&run.timestamp).map_or_else(
                     |_| run.timestamp.clone(),
                     |dt| dt.format("%Y-%m-%d %H:%M:%S").to_string(),
                 );
 
-                // Format command for display (truncate if too long)
                 let command_display = shlex::try_join(run.command.iter().map(String::as_str))
                     .unwrap_or_else(|_| run.command.join(" "));
                 let command_truncated = if command_display.len() > command_width {
@@ -681,7 +294,7 @@ path = \".\"
             let metadata_content = std::fs::read_to_string(&metadata_path).with_context(|| {
                 format!("Failed to read metadata from {}", metadata_path.display())
             })?;
-            let metadata: RunMetadata =
+            let metadata: capsula_orchestration::vault::RunMetadata =
                 serde_json::from_str(&metadata_content).with_context(|| {
                     format!("Failed to parse metadata from {}", metadata_path.display())
                 })?;
@@ -715,7 +328,6 @@ path = \".\"
                 anyhow::bail!("Either provide a run ID/name or use --all flag");
             }
 
-            // Priority: CLI flag > Environment variable (after dotenv) > Config file
             let server_url =
                 resolve_server_url(server, config.server.as_deref()).ok_or_else(|| {
                     anyhow::anyhow!(
@@ -723,11 +335,9 @@ path = \".\"
                     )
                 })?;
 
-            // Create client and check if vault exists
             let client = capsula_client::CapsulaClient::new(&server_url);
             let vault_name = &config.vault.name;
 
-            // Check if vault exists on server (only once)
             match client.vault_exists(vault_name) {
                 Ok(None) => {
                     warn!(
@@ -744,7 +354,6 @@ path = \".\"
             }
 
             if all {
-                // Push all runs in the vault
                 info!(
                     "Pushing all runs from vault '{}' to server {}",
                     vault_name, server_url
@@ -754,7 +363,6 @@ path = \".\"
                 let mut skip_count = 0;
                 let mut error_count = 0;
 
-                // Iterate through all run directories
                 for entry in walkdir::WalkDir::new(&vault_dir)
                     .min_depth(2)
                     .max_depth(2)
@@ -799,7 +407,6 @@ path = \".\"
                     anyhow::bail!("{error_count} runs failed to push");
                 }
             } else {
-                // Push single run
                 let run_id = run_id
                     .as_ref()
                     .expect("run_id must be Some when all is false");
@@ -813,7 +420,6 @@ path = \".\"
         }
         Commands::Vaults { command } => match command {
             VaultsCommands::List { server } => {
-                // Priority: CLI flag > Environment variable (after dotenv) > Config file
                 let server_url =
                     resolve_server_url(server, config.server.as_deref()).ok_or_else(|| {
                         anyhow::anyhow!(
