+++ crates/capsula-orchestration/Cargo.toml
@@ -12,13 +12,18 @@ categories = ["development-tools"]
 
 [dependencies]
 anyhow = { workspace = true }
+capsula-client = { workspace = true }
 capsula-config = { workspace = true }
 capsula-core = { workspace = true }
 capsula-registry = { workspace = true }
+chrono = { workspace = true }
 names = { workspace = true }
+reqwest = { workspace = true }
+serde = { workspace = true }
 serde_json = { workspace = true }
 tracing = { workspace = true }
 ulid = { workspace = true }
+walkdir = "2"
 
 [lints]
 workspace = true
