+++ crates/capsula-orchestration/src/lib.rs
@@ -1,2 +1,5 @@
 pub mod hooks;
+pub mod push;
+pub mod resolve;
 pub mod run;
+pub mod vault;
